Combine symlinks functions into MakeCommands
