+++ src/Command/MakeCommands.php
@@ -2,7 +2,7 @@
 
 namespace DkanTools\Command;
 
-use DkanTools\SymlinksTrait;
+use DkanTools\DrupalProjectTrait;
 use DkanTools\Util\Util;
 use Robo\Tasks;
 
@@ -13,7 +13,7 @@ use Robo\Tasks;
  */
 class MakeCommands extends Tasks
 {
-    use SymlinksTrait;
+    use DrupalProjectTrait;
 
     const DRUPAL_FOLDER_NAME = "docroot";
 
@@ -55,8 +55,62 @@ class MakeCommands extends Tasks
         $composerInstall->run();
 
         // Symlink dirs from src into docroot.
-        $this->addSymlinksToDrupalRoot();
+        $this->makeSymlinks();
 
         $this->io()->success("dktl make completed.");
     }
+
+    /**
+     * Create symlinks from docroot to folders in src.
+     *
+     * This will run automatically as part of the main make command, but can be
+     * run separately if your symlinks have been lost but you don't actually
+     * need to re-run make.
+     */
+    public function makeSymlinks()
+    {
+        $targetsAndLinks = [
+            ['target' => 'src/site',    'link' => '/sites/default'],
+            ['target' => 'src/modules', 'link' => '/modules/custom'],
+            ['target' => 'src/themes',  'link' => '/themes/custom'],
+            ['target' => 'src/schema',  'link' => '/schema'],
+        ];
+        foreach ($targetsAndLinks as $targetAndLink) {
+            $this->docrootSymlink(
+                $targetAndLink['target'],
+                self::$drupalDocroot . $targetAndLink['link']
+            );
+        }
+    }
+
+    private function docrootSymlink($target, $link)
+    {
+        $project_dir = Util::getProjectDirectory();
+        $target = $project_dir . "/{$target}";
+        $link = $project_dir . "/{$link}";
+        $link_parts = pathinfo($link);
+        $link_dirname = $link_parts['dirname'];
+        $target_path_relative_to_link = (new Filesystem())->makePathRelative($target, $link_dirname);
+
+        if (!file_exists($target) || !file_exists(self::$drupalDocroot)) {
+            $this->io()->warning(
+                "Skipping linking $target. Folders $target and '" .
+                self::$drupalDocroot . "' must both be present to create link."
+            );
+            return;
+        }
+
+        $result = $this->taskFilesystemStack()->stopOnFail()
+            ->remove($link)
+            ->symlink($target_path_relative_to_link, $link)
+            ->run();
+
+        if ($result->getExitCode() != 0) {
+            $this->io()->warning('Could not create link');
+        } else {
+            $this->io()->success("Symlinked $target to $link");
+        }
+        return $result;
+    }
+
 }
